# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MEETable API Docs",
    "description": "MEETable api 문서입니다",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "name": "authorization",
      "type": "apiKey",
      "in": "header"
    }
  },
...
  "paths": {
...
    "/home/bookmark": {
      "patch": {
        "tags": [
          "home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "isBookmark": {
                  "example": "T or F"
                },
                "promiseId": {
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home/folder": {
      "post": {
        "description": "not in use",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "foldereName": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home/totalpromise": {
      "get": {
        "tags": [
          "home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "description": "name: 이름 순 정렬, id: 생성일 순 정렬",
            "schema": {
              "type": "string"
              }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "access denied"
          },
          "500": {
            "description": "some errors"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home/deletepromise": {
      "patch": {
        "tags": [
          "home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "promiseId": {
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home/trash": {
      "get": {
        "tags": [
          "home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "description": "name: 이름 순 정렬, id: 생성일 순 정렬",
            "schema": {
              "type": "string"
              }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "access denied"
          },
          "500": {
            "description": "some errors"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home/promisename": {
      "patch": {
        "tags": [
          "home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "promiseName": {
                  "example": "약속22"
                },
                "promiseId": {
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home/search": {
      "get": {
        "tags": [
          "home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "description": "검색어를 넣어주세요",
            "schema": {
              "type": "string"
              }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "access denied"
          },
          "500": {
            "description": "some errors"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home/backoutpromise": {
      "delete": {
        "tags": [
          "home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "promiseId": {
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }, "400": {
            "description": "Bad Request"
          }, "500": {
            "description": "백엔드 에러"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}